+++ src/__temp/services/ProjectService/handlers/deleteProject.ts
@@ -1,5 +1,5 @@
-import DB from "@src/services/DBService";
-import ProjectService from "@src/services/ProjectService";
+import DB from "@src/__temp/services/DBService";
+import ProjectService from "@src/__temp/services/ProjectService";
 
 const deleteProject = async (
   service: ProjectService,
