+++ src/core/createSlice.ts
@@ -1,6 +1,6 @@
-import type { DomainAggregate } from "../domains/types/DomainAggregate";
+import type { DomainAggregate } from "./domain/types/DomainAggregate";
 import type { BaseHandler } from "./types/BaseHandler";
-import type { DomainEvent } from "../domains/types/DomainEvent";
+import type { DomainEvent } from "./domain/types/DomainEvent";
 import type { Slice } from "./types/Slice";
 
 export function createSlice<AggregateName extends DomainAggregate["name"]>({
