+++ src/core/types/Slice.ts
@@ -1,4 +1,4 @@
-import type { DomainEvent } from "../../domains/types/DomainEvent";
+import type { DomainEvent } from "../domain/types/DomainEvent";
 import type { BaseHandler } from "./BaseHandler";
 
 export type Slice = {
