--- src/services/DBService/index.ts
@@ -1,11 +1,0 @@
-import DeploymentService from "./deployment";
-import ProjectService from "./project";
-import UserService from "./user";
-
-class DB {
-  static User = UserService;
-  static Deployment = DeploymentService;
-  static Project = ProjectService;
-}
-
-export default DB;
