--- src/services/Logger/index.ts
@@ -1,157 +1,0 @@
-import fs from "fs/promises";
-import chalk from "chalk";
-
-import Config from "@src/config";
-import Observer from "@src/services/Observer";
-import getFormattedKoreaTime from "./utils/getFormattedKoreaTime";
-
-import { LogMessage, LogType } from "@src/types";
-
-/*
- * how to use
- *
- * import log from "@src/services/Logger";
- *
- * log.build(logType, message, [...message]);
- *   ㄴ> Logs from project deployment
- *
- * log.debug(logType, message, [...message]);
- *   ㄴ> server log (ex: connect server, morgan log, etc..)
- */
-
-class Logger extends Observer {
-  /* logging handlers */
-  private static console(logType: LogType, ...messages: LogMessage[]) {
-    const messageHead = `[${getFormattedKoreaTime(new Date())}]`;
-
-    switch (logType) {
-      case LogType.Server: {
-        for (const message of messages) {
-          process.stdout.write(
-            [
-              chalk.greenBright.bold("[COMMON]"),
-              chalk.blackBright(messageHead),
-              chalk.greenBright(message),
-            ].join(" ") + "\n",
-          );
-        }
-
-        break;
-      }
-      case LogType.Request: {
-        for (const message of messages) {
-          process.stdout.write(
-            [
-              chalk.magentaBright.bold("[REQUEST]"),
-              chalk.blackBright(messageHead),
-              chalk.magentaBright(message),
-            ].join(" ") + "\n",
-          );
-        }
-
-        break;
-      }
-      case LogType.Deployment: {
-        for (const message of messages) {
-          process.stdout.write(
-            [
-              chalk.hex("#EE9560").bold("[BUILD]"),
-              chalk.blackBright(messageHead),
-              chalk.hex("#EE9560")(message),
-            ].join(" ") + "\n",
-          );
-        }
-
-        break;
-      }
-      case LogType.Error: {
-        for (const message of messages) {
-          process.stderr.write(
-            [
-              chalk.redBright.bold("[ERROR]"),
-              chalk.redBright(messageHead),
-              chalk.redBright.bold(message),
-            ].join(" ") + "\n",
-          );
-        }
-
-        break;
-      }
-      default: {
-        break;
-      }
-    }
-  }
-
-  private static async writefile(logType: LogType, ...messages: LogMessage[]) {
-    const messageHead = `[${getFormattedKoreaTime(new Date())}]`;
-
-    switch (logType) {
-      case LogType.Server:
-      case LogType.Request: {
-        for (const message of messages) {
-          await fs.appendFile(
-            `logs/server.log`,
-            `${messageHead} ${message} \n`,
-          );
-        }
-      }
-      case LogType.Deployment: {
-        for (const message of messages) {
-          await fs.appendFile(
-            `logs/deployment.log`,
-            `${messageHead} ${message} \n`,
-          );
-        }
-      }
-      case LogType.Error: {
-        for (const message of messages) {
-          await fs.appendFile(`logs/error.log`, `${messageHead} ${message} \n`);
-        }
-      }
-      default: {
-        break;
-      }
-    }
-  }
-
-  private static notify(...messages: LogMessage[]) {
-    const messageHead = `[${getFormattedKoreaTime(new Date())}]`;
-
-    for (const message of messages) {
-      Logger.send(`${messageHead} ${message}`);
-    }
-  }
-
-  /* logging methods */
-  static build(...messages: LogMessage[]) {
-    if (Config.CLIENT_OPTIONS.debug) {
-      Logger.console(LogType.Deployment, ...messages);
-    }
-
-    Logger.writefile(LogType.Deployment, ...messages);
-    Logger.notify(...messages);
-  }
-
-  static buildError(...messages: LogMessage[]) {
-    Logger.writefile(LogType.Error, ...messages);
-    Logger.writefile(LogType.Deployment, ...messages);
-    Logger.notify(...messages);
-  }
-
-  static debug(...messages: LogMessage[]) {
-    Logger.writefile(LogType.Server, ...messages);
-    Logger.console(LogType.Server, ...messages);
-  }
-
-  static serverError(...messages: LogMessage[]) {
-    Logger.writefile(LogType.Error, ...messages);
-    Logger.console(LogType.Error, ...messages);
-  }
-
-  static request(...messages: LogMessage[]) {
-    Logger.console(LogType.Request, ...messages);
-  }
-}
-
-export default Logger;
